requirements.txt
Gunicorn
antiorm==1.2.1
attrs==22.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
fonttools==4.38.0
gradescope-utils==0.5.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
login==0.0.6
MarkupSafe==2.1.2
matplotlib==3.7.0
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
protobuf==4.22.0
pygame==2.1.3
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.19.0
Werkzeug==2.2.2